Change and check coding

+++ moneytracker-backend/src/main/java/com/example/moneytrackerbackend/services/CategoryServiceImp.java
@@ -8,24 +8,28 @@ import com.example.moneytrackerbackend.exceptiones.CustomException;
 import com.example.moneytrackerbackend.repositories.CategoryRepository;
 import com.example.moneytrackerbackend.repositories.TransactionRepository;
 import com.example.moneytrackerbackend.repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class CategoryServiceImp implements CategoryService{
-    @Autowired
-    private CategoryRepository categoryRepository;
-    @Autowired
-    private TransactionRepository transactionRepository;
-    @Autowired
-    private UserRepository userRepository;
-    public Category createCategory(CategoryRequest categoryRequest){
-        if(categoryRepository.existsByTitle(categoryRequest.getTitle(), categoryRequest.getUserId()) >0){
-            throw new CustomException("Error: Name category da ton tai!!!");
+
+public class CategoryServiceImp implements CategoryService {
+    private final CategoryRepository categoryRepository;
+    private final TransactionRepository transactionRepository;
+    private final UserRepository userRepository;
+
+    public CategoryServiceImp(CategoryRepository categoryRepository, TransactionRepository transactionRepository, UserRepository userRepository) {
+        this.categoryRepository = categoryRepository;
+        this.transactionRepository = transactionRepository;
+        this.userRepository = userRepository;
+    }
+
+    public Category createCategory(CategoryRequest categoryRequest) {
+        if (categoryRepository.existsByTitle(categoryRequest.getTitle(), categoryRequest.getUserId()) > 0) {
+            throw new CustomException("Error: Name category already exists!!!");
         }
-        User user = userRepository.findById(categoryRequest.getUserId()).orElseThrow(()-> new CustomException("Error: no use"));
+        User user = userRepository.findById(categoryRequest.getUserId()).orElseThrow(() -> new CustomException("Error: no use"));
 
         Category category = Category.builder()
                 .title(categoryRequest.getTitle())
@@ -33,27 +37,25 @@ public class CategoryServiceImp implements CategoryService{
                 .value(categoryRequest.isValue()).build();
         return categoryRepository.save(category);
     }
-    public List<Category> getAllCategory(Long userId){
-        User user = userRepository.findById(userId).orElseThrow(()-> new CustomException("Error: no use"));
-        List<Category> categories=categoryRepository.findAllByUser(user);
-        return categories;
-    }
-    public List<Category> getAllByValue(boolean value, Long userId){
-        User user = userRepository.findById(userId).orElseThrow(()-> new CustomException("Error: no use"));
-        List<Category> categories = categoryRepository.findAllByValueAndUser(value,user);
-        return categories;
+
+    public List<Category> getAllCategory(Long userId) {
+        return categoryRepository.findAllByUserId(userId);
     }
 
-    public Category getCategoryById(Long id){
-        Category category = categoryRepository.findById(id)
-                .orElseThrow(() -> new CustomException("Error: no category"));
-        return category;
+    public List<Category> getAllByValue(boolean value, Long userId) {
+        return categoryRepository.findAllByValueAndUserId(value, userId);
     }
-    public void deleteCategory(Long id){
+
+    public Category getCategoryById(Long id) {
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new CustomException("Error: no category"));
+    }
+
+    public void deleteCategory(Long id) {
         Category category = getCategoryById(id);
-        List<Transaction> transactions = transactionRepository.findAllByCategoryOrderByDate(category);
-        for(Transaction transaction: transactions){
-//            transaction.setCategory();
+        List<Transaction> transactions = transactionRepository.findAllByCategoryIdOrderByDate(category.getId());
+        if(!transactions.isEmpty()){
+            throw new CustomException("Error: Have transaction of this category");
         }
         categoryRepository.deleteById(id);
     }
